--- frontend/src/stylesheets/profile.css
@@ -1,12 +1,0 @@
-.root{
-    margin-left: 25%;
-    width: 50%;
-}
-
-.title{
-    font-size: 80,
-  }
-
-.pos{
-    margin-bottom: 12,
-}
